+++ src/models/index.ts
@@ -1,5 +1,6 @@
 export * from './Choice';
 export * from './ContentFolder';
+export * from './CustomTaxonomyData';
 export * from './DashboardData';
 export * from './DraftField';
 export * from './Framework';