+++ src/panelWebView/components/TagPicker.tsx
@@ -7,9 +7,10 @@ import Downshift from 'downshift';
 import { AddIcon } from './Icons/AddIcon';
 import { VsLabel } from './VscodeComponents';
 import { MessageHelper } from '../../helpers/MessageHelper';
+import { CustomTaxonomyData } from '../../models';
 
 export interface ITagPickerProps {
-  type: string;
+  type: TagType;
   icon: JSX.Element;
   label?: string;
   crntSelected: string[];
@@ -18,10 +19,12 @@ export interface ITagPickerProps {
   focussed: boolean;
   unsetFocus: () => void;
   disableConfigurable?: boolean;
+  fieldName?: string;
+  taxonomyId?: string;
 }
 
 const TagPicker: React.FunctionComponent<ITagPickerProps> = (props: React.PropsWithChildren<ITagPickerProps>) => {
-  const { label, icon, type, crntSelected, options, freeform, focussed, unsetFocus, disableConfigurable } = props;
+  const { label, icon, type, crntSelected, options, freeform, focussed, unsetFocus, disableConfigurable, fieldName, taxonomyId } = props;
   const [ selected, setSelected ] = React.useState<string[]>([]);
   const [ inputValue, setInputValue ] = React.useState<string>("");
   const prevSelected = usePrevious(crntSelected);
@@ -43,26 +46,37 @@ const TagPicker: React.FunctionComponent<ITagPickerProps> = (props: React.PropsW
    * @param tag 
    */
   const onCreate = (tag: string) => {
-    const cmdType = type === TagType.tags ? CommandToCode.addTagToSettings : CommandToCode.addCategoryToSettings;
-    MessageHelper.sendMessage(cmdType, tag);
+    if (type === TagType.tags) {
+      MessageHelper.sendMessage(CommandToCode.addTagToSettings, tag);
+    } else if (type === TagType.categories) {
+      MessageHelper.sendMessage(CommandToCode.addCategoryToSettings, tag);
+    } else if (type === TagType.custom) {
+      MessageHelper.sendMessage(CommandToCode.addToCustomTaxonomy, {
+        id: taxonomyId,
+        name: fieldName,
+        option: tag
+      } as CustomTaxonomyData);
+    }
   };
 
   /**
    * Send an update to VSCode
    * @param values 
    */
-  const sendUpdate = (values: string[]) => {
-    let cmdType = CommandToCode.updateCategories;
-    
+  const sendUpdate = (values: string[]) => {    
     if (type === TagType.tags) {
-      cmdType = CommandToCode.updateTags;
+      MessageHelper.sendMessage(CommandToCode.updateTags, values);
     } else if (type === TagType.categories) {
-      cmdType = CommandToCode.updateCategories;
+      MessageHelper.sendMessage(CommandToCode.updateCategories, values);
     } else if (type === TagType.keywords) {
-      cmdType = CommandToCode.updateKeywords;
+      MessageHelper.sendMessage(CommandToCode.updateKeywords, values);
+    } else if (type === TagType.custom) {
+      MessageHelper.sendMessage(CommandToCode.updateCustomTaxonomy, {
+        id: taxonomyId,
+        name: fieldName,
+        options: values
+      } as CustomTaxonomyData);
     }
-
-    MessageHelper.sendMessage(cmdType, values);
   };
 
   /**